+++ CHANGELOG.md
@@ -5,9 +5,11 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 - [POSSIBLE BC BREAK] moved main configurations from page root to layout (cause this is the please where it belongs)! While we provide a runonce that do some migration, we strongly recommend to check your configuration. Also no entries are moved to the layout (which issn't breaking, cause they still are loaded from page since entry overriding not changed.
+- don't use global $objPage in generatePage hook
 
 ### Fixed
 - possible error on fresh installation
+- added missing utils bundle dependency
 
 
 ## [0.8.3] - 2019-04-23